+++ src/BethesdaMultitool/Core/Formats/Nif/Rendering/Npc/NpcBoundaryVertexStitcher.cs
@@ -1,4 +1,5 @@
 using BethesdaMultitool.Core.Diagnostics;
+
 namespace BethesdaMultitool.Core.Formats.Nif.Rendering.Npc;
 
 /// <summary>
@@ -199,7 +200,7 @@ internal static class NpcBoundaryVertexStitcher
         // Pack three 21-bit integers into a 64-bit key
         return ((long)(ix & 0x1FFFFF) << 42) |
                ((long)(iy & 0x1FFFFF) << 21) |
-               (long)(iz & 0x1FFFFF);
+               iz & 0x1FFFFF;
     }
 
     private static void ClearBindPoseData(List<RenderableSubmesh> submeshes)
